Final commit

+++ index.js
@@ -5,6 +5,7 @@ const mongoose = require("mongoose");
 const { engine } = require("express-handlebars");
 const jwt = require("jsonwebtoken");
 const bcrypt = require("bcryptjs");
+const Joi = require("joi");
 
 const app = express();
 const bodyParser = require("body-parser"); // pull information from HTML POST (express4)
@@ -48,6 +49,45 @@ function verifyToken(req, res, next) {
   }
 }
 
+// Schema definitions for object validations during api calls
+const userSignUpSchema = Joi.object({
+  firstName: Joi.string().max(20).required(),
+  lastName: Joi.string().max(20).required(),
+  username: Joi.string().alphanum().min(3).max(20).required(),
+  password: Joi.string(),
+});
+
+const userLoginSchema = Joi.object({
+  username: Joi.string().alphanum().min(3).max(20).required(),
+  password: Joi.string(),
+});
+
+const restaurantBodySchema = Joi.object({
+  address: Joi.object({
+    building: Joi.string().required(),
+    coord: Joi.array().items(Joi.number()).min(2).max(2).required(),
+    street: Joi.string().required(),
+    zipcode: Joi.string().min(5).max(10).required(),
+  }),
+  borough: Joi.string().required(),
+  cuisine: Joi.string().required(),
+  grades: Joi.array()
+    .items(
+      Joi.object({
+        date: Joi.date().iso().required(),
+        grade: Joi.string().valid("A", "B", "C", "D", "E").required(),
+        score: Joi.number().integer().min(0).max(100).required(),
+      })
+    )
+    .optional(),
+  name: Joi.string().required(),
+  restaurant_id: Joi.string().required(),
+});
+
+const validateId = Joi.object({
+  id: Joi.string().required(),
+});
+
 // Route for signup form
 app.get("/signup", (req, res) => {
   res.status(200).render("signup");
@@ -65,16 +105,22 @@ app.post("/signup", async (req, res) => {
     password: password,
   };
 
-  db.addNewUser(user)
-    .then(() => {
-      const accessToken = jwt.sign(user, process.env.JWT_KEY);
-      res
-        .status(200)
-        .json({ info: "Signed Up Successfully", accessToken: accessToken });
-    })
-    .catch((error) => {
-      res.status(500).json(error);
-    });
+  const parseRes = userSignUpSchema.validate(user);
+
+  if (parseRes.error === undefined) {
+    db.addNewUser(user)
+      .then(() => {
+        const accessToken = jwt.sign(user, process.env.JWT_KEY);
+        res
+          .status(200)
+          .json({ info: "Signed Up Successfully", accessToken: accessToken });
+      })
+      .catch((error) => {
+        res.status(500).json(error);
+      });
+  } else {
+    res.status(422).json({ info: "Invalid request body params!" });
+  }
 });
 
 // Route for login form
@@ -86,16 +132,22 @@ app.get("/login", (req, res) => {
 app.post("/login", (req, res) => {
   const user = { username: req.body.username, password: req.body.password };
 
-  db.isValidUser(user).then((response) => {
-    if (response) {
-      const accessToken = jwt.sign(user, process.env.JWT_KEY);
-      res
-        .status(200)
-        .json({ info: "Login Successful", accessToken: accessToken });
-    } else {
-      res.status(404).json({ info: "User not found" });
-    }
-  });
+  const parseRes = userLoginSchema.validate(user);
+
+  if (parseRes.error === undefined) {
+    db.isValidUser(user).then((response) => {
+      if (response) {
+        const accessToken = jwt.sign(user, process.env.JWT_KEY);
+        res
+          .status(200)
+          .json({ info: "Login Successful", accessToken: accessToken });
+      } else {
+        res.status(404).json({ info: "User not found" });
+      }
+    });
+  } else {
+    res.status(422).json({ info: "Invalid request body params!" });
+  }
 });
 
 // Route to display the form
@@ -105,13 +157,21 @@ app.get("/api/restaurants/form", (req, res) => {
 
 // Route to add new restaurant to db
 app.post("/api/restaurants", verifyToken, (req, res) => {
-  db.addNewRestaurant(req.body)
-    .then((response) => {
-      res.status(200).json({ info: "Restaurant Data Inserted", res: response });
-    })
-    .catch((error) => {
-      res.status(500).json(error);
-    });
+  const parseRes = restaurantBodySchema.validate({ ...req.body });
+  console.log(parseRes);
+  if (parseRes.error === undefined) {
+    db.addNewRestaurant(req.body)
+      .then((response) => {
+        res
+          .status(200)
+          .json({ info: "Restaurant Data Inserted", res: response });
+      })
+      .catch((error) => {
+        res.status(500).json(error);
+      });
+  } else {
+    res.status(422).json({ info: "Invalid request body params!" });
+  }
 });
 
 // Route to get all restaurants from db
@@ -155,26 +215,40 @@ app.get("/api/restaurants/:id", (req, res) => {
 
 // Route to Update Restaurant By Id
 app.put("/api/restaurants/:id", verifyToken, (req, res) => {
-  db.updateRestaurantById(req.params.id, req.body)
-    .then((response) => {
-      res.status(200).json({ info: "Restaurant data Updated", res: response });
-    })
-    .catch((error) => {
-      res.status(500).json(error);
-    });
+  const parseRes = restaurantBodySchema.validate({ ...req.body });
+  const parseId = Joi.string(req.params.id).required();
+
+  if (parseRes.error === undefined && parseId.error === undefined) {
+    db.updateRestaurantById(req.params.id, req.body)
+      .then((response) => {
+        res
+          .status(200)
+          .json({ info: "Restaurant data Updated", res: response });
+      })
+      .catch((error) => {
+        res.status(500).json(error);
+      });
+  } else {
+    res.status(422).json({ info: "Invalid request id or body params!" });
+  }
 });
 
 // Route to Delete Restaurant By Id
 app.delete("/api/restaurants/:id", verifyToken, (req, res) => {
-  db.deleteRestaurantById(req.params.id)
-    .then((response) => {
-      res
-        .status(200)
-        .json({ info: "Restaurant Deleted from DB", res: response });
-    })
-    .catch((error) => {
-      res.status(500).json(error);
-    });
+  const parseId = validateId.validate({ id: req.params.id });
+  if (parseId.error === undefined) {
+    db.deleteRestaurantById(req.params.id)
+      .then((response) => {
+        res
+          .status(200)
+          .json({ info: "Restaurant Deleted from DB", res: response });
+      })
+      .catch((error) => {
+        res.status(500).json(error);
+      });
+  } else {
+    res.status(422).json({ info: "Invalid request id or body params!" });
+  }
 });
 
 // Route to handle form submission and display results
@@ -209,3 +283,5 @@ mongoose
   .catch((error) => {
     console.log("DB Connection FAILED!", error);
   });
+
+module.exports = app;
